--- src/components/layouts/Tabordion/index.js
@@ -1,3 +1,0 @@
-export { default as Tabordion } from './Tabordion'
-export { Tab } from './Tabordion'
-export { TabPanel } from './Tabordion'
