+++ src/components/layouts/Tabtastic/index.js
@@ -1,0 +1,2 @@
+export { default as Tabtastic } from './Tabtastic'
+export { TabtasticTab } from './Tabtastic'
