+++ src/pages/index.mdx
@@ -15,9 +15,10 @@ import { makeStyles } from '@material-ui/core/styles'
 import Container from '@material-ui/core/Container'
 import Box from '@material-ui/core/Box'
 import Grid from '@material-ui/core/Grid'
+import Typography from '@material-ui/core/Typography'
 
 import { HomeHero } from '../components/sections/HomeHero'
-import { Tabordion, Tab, TabPanel } from '../components/layouts/Tabordion'
+import { Tabtastic, TabtasticTab } from '../components/layouts/Tabtastic'
 import { WhatsNew } from '../components/sections/WhatsNew'
 
 import TotalRevenue from '../components/sections/TotalRevenue'
@@ -30,16 +31,6 @@ import ExploreProduction from '../components/sections/Explore/Production'
 import { RevenueByLocationLink } from '../components/layouts/RevenueByLocationLink'
 
 export const useStyles = makeStyles(theme => ({
-  revenueByLocation: {
-    position: 'relative',
-    top: 0,
-  },
-  tabordionContainer: {
-    marginTop: -50,
-    '& .MuiTabs-fixed': {
-      marginRight: '350px',
-    }
-  },
   whatsNewContainer: {
     background: theme.palette.blue.light,
   }
@@ -79,18 +70,23 @@ export const IndexPage = props => {
       </Container>
       {/* Tabordion */}
       <Container maxWidth="lg">
-        <Box mb={15}>
-          <Grid container spacing={0} styles={{ 'padding-bottom': 0 }}>
-            <Grid item xs={12} md={12} className={classes.revenueByLocation}>
-              <RevenueByLocationLink />
-            </Grid>
-            <Grid item xs={12} md={12} className={classes.tabordionContainer}>
-              <Tabordion selected={selected}>
-                <Tab label="Revenue" id="tab-revenue" />
-                <Tab label="Disbursements" id="tab-disbursements" />
-                <Tab label="Production" id="tab-production" />
-                <TabPanel id="tab-panel-revenue">
+        <Box mb={5}>
+          <Grid container spacing={0}>
+            <Grid item xs={12} md={12}>
+              <Tabtastic selected={selected}>
+                <TabtasticTab label="Revenue">
                   <Grid container spacing={4}>
+                    <Grid item xs={9}>
+                      <Typography variant="h2">
+                        Revenue
+                      </Typography>
+                      <Typography variant="body1">
+                        The amount of money collected by the federal government from energy and mineral extraction on federal lands and waters and Native American lands.
+                      </Typography>
+                    </Grid>
+                    <Grid item xs={3}>
+                      <RevenueByLocationLink />
+                    </Grid>
                     <Grid item xs={12} md={8}>
                       Total revenue content...
                     </Grid>
@@ -101,9 +97,17 @@ export const IndexPage = props => {
                       <ExploreRevenue />
                     </Grid>
                   </Grid>
-                </TabPanel>
-                <TabPanel id="tab-panel-disbursements">
+                </TabtasticTab>
+                <TabtasticTab label="Disbursements">
                   <Grid container spacing={4}>
+                    <Grid item xs={12}>
+                      <Typography variant="h2">
+                        Disbursements
+                      </Typography>
+                      <Typography variant="body1">
+                        The amount of money the federal government distributed to various funds, agencies, local governments, and Native Americans.
+                      </Typography>
+                    </Grid>
                     <Grid item xs={12} md={8}>
                       Disbursement content...
                     </Grid>
@@ -114,9 +118,17 @@ export const IndexPage = props => {
                       <ExploreDisbursements />
                     </Grid>
                   </Grid>
-                </TabPanel>
-                <TabPanel id="tab-panel-production">
+                </TabtasticTab>
+                <TabtasticTab label="Production">
                   <Grid container spacing={4}>
+                    <Grid item xs={12}>
+                      <Typography variant="h2">
+                        Production
+                      </Typography>
+                      <Typography variant="body1">
+                        The volume of major commodities extracted on federal lands and waters and Native American lands.
+                      </Typography>
+                    </Grid>
                     <Grid item xs={12}>
                       Production content...
                     </Grid>
@@ -124,8 +136,8 @@ export const IndexPage = props => {
                       <ExploreProduction />
                     </Grid>
                   </Grid>
-                </TabPanel>
-              </Tabordion>
+                </TabtasticTab>
+              </Tabtastic>
             </Grid>
           </Grid>
         </Box>
